requirements.txt
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
Flask==2.2.3
Flask-Cors==3.0.10
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==5.5.2
